#1848, move `DiscordenoUser` and `transformUser` from `src/transformers/member.ts` to `/src/transformers/user.ts`
#1848 Document audit logs for scheduled events

move `DiscordenoUser` interface and `transformUser` function from `src/transformers/member.ts` to `/src/transformers/user.ts`

add `export * from "./user.ts"` to /src/transformers/mod.ts
