+++ src/helpers/guilds/scheduledEvents/editScheduledEvent.ts
@@ -32,6 +32,7 @@ export async function editScheduledEvent(
       privacy_level: options.privacyLevel,
       entity_type: options.entityType,
       status: options.status,
+      reason: options.reason,
     }
   );
